🚨 small linter and format improvements

+++ test/algorithms/test_qft.cpp
@@ -87,7 +87,7 @@ TEST_P(QFT, Functionality) {
   // since only positive real values are stored in the complex table
   // this number has to be divided by 4
   ASSERT_EQ(dd->cn.realCount(),
-            static_cast<unsigned int>(std::ceil(std::pow(2, nqubits) / 4)));
+            static_cast<std::size_t>(std::ceil(std::pow(2, nqubits) / 4)));
 
   // top edge weight should equal sqrt(0.5)^n
   EXPECT_NEAR(dd::RealNumber::val(func.w.r),
@@ -128,7 +128,7 @@ TEST_P(QFT, FunctionalityRecursive) {
   // since only positive real values are stored in the complex table
   // this number has to be divided by 4
   ASSERT_EQ(dd->cn.realCount(),
-            static_cast<unsigned int>(std::ceil(std::pow(2, nqubits) / 4)));
+            static_cast<std::size_t>(std::ceil(std::pow(2, nqubits) / 4)));
 
   // top edge weight should equal sqrt(0.5)^n
   EXPECT_NEAR(dd::RealNumber::val(func.w.r),
